+++ doc/change-log.md
@@ -1,5 +1,6 @@
 # 3.4.1 - [2022-08-18]
 - `com.lucendar.strm.common.StreamingApi.STRM_FORMAT__RTSP` added
+- `com.lucendar.strm.common.StrmMediaNotif.CLOSE_CAUSE__termCmdError` added
 
 # 3.4.0 - [2022-05-12]
 ## Changed